fabzclean-fastapi/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
PyMySQL==1.1.0
alembic==1.11.1
python-dotenv==1.0.0
argon2-cffi==21.3.0
python-jose==3.3.0
qrcode==7.4.2
Pillow==10.0.0
psutil==5.9.5
pytest==7.4.2

